Add sidebar checker holder functionality and vertical checkers
As of now, if a checker is manually moved from the sidebar holder,
the checker selected doesn't have to match the color of the last vertical checker,
so the colors of the vertical checkers do not match those that they should be.
As manual checker moving from the vertical holder will be disabled,
it is not a concern.

The regular checkers are hidden behind the holder,
they are not removed so that undos can be easily performed.